+++ lib/mini_form/errors.rb
@@ -1,4 +1,11 @@
 module MiniForm
   class InvalidForm < StandardError
+    attr_reader :errors
+
+    def initialize(object)
+      @errors = object.errors
+
+      super "Form validation failed for: #{errors.keys.join(', ')}"
+    end
   end
 end