+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Version 0.2.0
 
-* Make `.attribute_names` public
+* Exposed `errors` on MiniForm::InvalidForm
+* Added descriptive message MiniForm::InvalidForm
+* Made `.attribute_names` public
 
 ## Version 0.1.0