+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.2.0
 
+* Added `assign_attributes` alias to `attributes=`
 * Exposed `errors` on MiniForm::InvalidForm
 * Added descriptive message MiniForm::InvalidForm
 * Made `.attribute_names` public