+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Version 0.2.0
 
+* Added read option to `model`:
+
+```ruby
+class EditProfile
+  include MiniForm::Model
+
+  model :user, attributes: %i(email name), read:%(id)
+end
+
+profile = EditProfile.new(user: user)
+profile.id
+profile.id = 1 # raises NoMethodError
+```
+
+
 * MiniForm::Model can be inherited
 * Added `assigment` callbacks, called when attributes are assigned
 * Added `assign_attributes` alias to `attributes=`